+++ x/gamm/keeper/genesis.go
@@ -7,11 +7,19 @@ import (
 	"github.com/osmosis-labs/osmosis/v10/x/gamm/types"
 )
 
+<<<<<<< HEAD
 // InitGenesis initializes the capability module's state from a provided genesis
 // state.
 func (k Keeper) InitGenesis(ctx sdk.Context, genState types.GenesisState, unpacker codectypes.AnyUnpacker) {
 	k.SetParams(ctx, genState.Params)
 	k.SetNextPoolNumber(ctx, genState.NextPoolNumber)
+=======
+// InitGenesis initializes the x/gamm module's state from a provided genesis
+// state, which includes the current live pools, global pool parameters (e.g. pool creation fee), next pool id etc.
+func (k Keeper) InitGenesis(ctx sdk.Context, genState types.GenesisState, unpacker codectypes.AnyUnpacker) {
+	k.setParams(ctx, genState.Params)
+	k.setNextPoolId(ctx, genState.NextPoolNumber)
+>>>>>>> 8cac0906 (gamm keeper delete obsolete files (#2160))
 
 	liquidity := sdk.Coins{}
 	for _, any := range genState.Pools {
@@ -49,7 +57,11 @@ func (k Keeper) ExportGenesis(ctx sdk.Context) *types.GenesisState {
 		poolAnys = append(poolAnys, any)
 	}
 	return &types.GenesisState{
+<<<<<<< HEAD
 		NextPoolNumber: k.GetNextPoolNumberAndIncrement(ctx),
+=======
+		NextPoolNumber: k.GetNextPoolId(ctx),
+>>>>>>> 8cac0906 (gamm keeper delete obsolete files (#2160))
 		Pools:          poolAnys,
 		Params:         k.GetParams(ctx),
 	}